Update configuration/docs & add integration tests
The configuration has been completely reworked using [`Dry::Configurable`](https://dry-rb.org/gems/dry-configurable/). This is a *huge* win over our previous `CryptIdent::Config` class and, more specifically, on the Ruby Goldberg gyrations we did to make `#configure_crypt_ident` et al kinda-sorta-usually work for a sufficiently loose definition of _work_ that it isn't even funny.

It also lets us take a look at both the main `CryptIdent` module's methods and the supporting classes and see what parameters we use (or don't). Specifically, the app we're developing this for can reasonably be expected never to call any of the methods with a non-default Repository specification. If that's needed, switching around the `CryptIdent.config.repository` value is straightforward enough and lets us eliminate a major source of fragility in tests.

Integration tests serve both as an important functionality check and as "how-to" documentation exercising various features and workflows for `CryptIdent`. The first such test developed, `test/integration/register_and_authenticate_test.rb`, includes comments detailing the rationale for these tests; see that file for more.

The tests make use of Hanami migrations (required for live `Hanami::Repository` instances). The `create_users` migration defines the columns we care about here: `password_hash`, `password_reset_expires_at`, and `token`. Not coincidentally, we've learned quite a bit about what (not) to do when bringing up a stunt-double Repository, such as moving the registration/migration logic into a file `require`d by multiple spec files, because copying/pasting the code in each results in errors due to duplicate initialisation (that *must* be performed exactly and only once.)

These tests also allowed us to discover and fix bugs that had slipped through unit tests. An example is in `#change_password`, which was generating the new Password data to be stored *completely and egregiously* incorrectly.

They each make use of several `CryptIdent` methods to demonstrate a "typical" sequence of Registering a New User, performing a Password Reset on that New User, Signing In, and Signing Out. The integration tests considered to most authoritatively document and exercise each `CryptIdent` method (besides `CryptIdent#config`, which is used in each), are:

| Method | Integration Test File |
|:------ |:--------------------- |
| `#change_password` | `test/integration/change_password_test.rb` |
| `#generate_reset_token` | **TBD** |
| `#reset_password` | `test/integration/register_and_authenticate_test.rb` |
| `#session_expired?` | `test/integration/session_expiry_test.rb` |
| `#sign_in` | `test/integration/register_and_authenticate_test.rb` |
| `#sign_out` | `test/integration/register_and_authenticate_test.rb` |
| `#sign_up` | `test/integration/register_and_authenticate_test.rb` |
| `#update_session_expiry` | `test/integration/demo_update_session_expiry_test.rb` |

    78 tests, 95 assertions, 0 failures, 0 errors, 0 skips
    Coverage: 987 / 987 LOC (100.0%) covered.
    RuboCop: 11 files inspected, no offenses detected
    Flay: Total score 0
    Flog: Total 310.7; method average 4.5; max 9.5 (CryptIdent::SignIn#call)
    Reek: 0 total warnings
    Inch: Nothing to suggest